+++ src/services/__tests__/templateService.spec.ts
@@ -1,7 +1,10 @@
 import casual from 'casual';
 import { Template, TemplateVisibility } from "../../models/Template";
 import { VersionedTemplate, TemplateVersionType } from '../../models/VersionedTemplate';
-import { cloneTemplate, generateTemplateVersion, hasPermissionOnTemplate } from '../templateService';
+import {
+  cloneTemplate, generateTemplateVersion, hasPermissionOnTemplate,
+  setDefaultTemplate
+} from '../templateService';
 import { TemplateCollaborator } from '../../models/Collaborator';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { isSuperAdmin } from '../authService';
@@ -448,5 +451,90 @@ describe('template versioning', () => {
     expect(updated.modified).toEqual(tmplt.modified);
     expect(updated.latestPublishVersion).toEqual(newVersion.version);
     expect(updated.isDirty).toEqual(true);
+  });
+});
+
+describe('setDefaultTemplate', () => {
+  let template: Template;
+  let oldTemplate: VersionedTemplate;
+
+  beforeEach(async () => {
+    jest.clearAllMocks();
+
+    template = new Template({
+      id: casual.integer(1, 99),
+      name: casual.sentence,
+      ownerId: casual.url,
+      createdById: casual.integer(1, 99),
+    });
+
+    oldTemplate = new VersionedTemplate({
+      id: casual.integer(100, 999),
+      templateId: casual.integer(1000, 9999),
+      name: casual.sentence,
+      ownerId: casual.url,
+      createdById: casual.integer(1, 99),
+    });
+  });
+
+  it('successfully marks the template as default when there was no default before', async () => {
+    jest.spyOn(VersionedTemplate, 'defaultTemplate').mockResolvedValue(undefined);
+    jest.spyOn(Template,'query').mockResolvedValue([{}]);
+    jest.spyOn(VersionedTemplate,'query').mockResolvedValue([{}]);
+    await setDefaultTemplate('Test', context, template);
+    expect(VersionedTemplate.defaultTemplate).toHaveBeenCalled();
+    expect(Template.query).toHaveBeenCalledTimes(1);
+    expect(VersionedTemplate.query).toHaveBeenCalledTimes(1);
+    expect(Template.query).toHaveBeenLastCalledWith(context, 'UPDATE templates SET isDefault = ? WHERE id = ?;', ['1', template.id.toString()], 'Test');
+    expect(VersionedTemplate.query).toHaveBeenLastCalledWith(context, 'UPDATE versionedTemplates SET isDefault = ? WHERE templateId = ?;', ['1', template.id.toString()], 'Test');
+  });
+
+  it('successfully marks the template as default when one already is set', async () => {
+    jest.spyOn(VersionedTemplate, 'defaultTemplate').mockResolvedValue(oldTemplate);
+    jest.spyOn(Template,'query').mockResolvedValue([{}]);
+    jest.spyOn(VersionedTemplate,'query').mockResolvedValue([{}]);
+    await setDefaultTemplate('Test', context, template);
+    expect(VersionedTemplate.defaultTemplate).toHaveBeenCalled();
+    expect(Template.query).toHaveBeenCalledTimes(2);
+    expect(VersionedTemplate.query).toHaveBeenCalledTimes(2);
+    expect(Template.query).toHaveBeenLastCalledWith(context, 'UPDATE templates SET isDefault = ? WHERE id = ?;', ['0', oldTemplate.templateId.toString()], 'Test');
+    expect(VersionedTemplate.query).toHaveBeenLastCalledWith(context, 'UPDATE versionedTemplates SET isDefault = ? WHERE templateId = ?;', ['0', oldTemplate.templateId.toString()], 'Test');
+  });
+
+  it('does not unmark the existing template as default if the marking failed', async () => {
+    jest.spyOn(VersionedTemplate, 'defaultTemplate').mockResolvedValue(undefined);
+    jest.spyOn(Template,'query').mockResolvedValue(undefined);
+    await setDefaultTemplate('Test', context, template);
+    expect(Template.query).toHaveBeenCalledTimes(1);
+    expect(Template.query).toHaveBeenLastCalledWith(context, 'UPDATE templates SET isDefault = ? WHERE id = ?;', ['1', template.id.toString()], 'Test');
+  });
+
+  it('rolls back if marking the versionedTemplates as default fails', async () => {
+    jest.spyOn(VersionedTemplate, 'defaultTemplate').mockResolvedValue(undefined);
+    jest.spyOn(Template,'query').mockResolvedValue([{}]);
+    jest.spyOn(VersionedTemplate,'query').mockResolvedValue(undefined);
+    await setDefaultTemplate('Test', context, template);
+    expect(Template.query).toHaveBeenCalledTimes(2);
+    expect(Template.query).toHaveBeenLastCalledWith(context, 'UPDATE templates SET isDefault = ? WHERE id = ?;', ['0', template.id.toString()], 'Test');
+  });
+
+  it('rolls back if the unmarking fails', async () => {
+    jest.spyOn(VersionedTemplate, 'defaultTemplate').mockResolvedValue(oldTemplate);
+    const tSpy = jest.spyOn(Template, 'query');
+    const vtSpy = jest.spyOn(VersionedTemplate, 'query');
+    tSpy.mockResolvedValueOnce([{}]); // Mark new template as default
+    vtSpy.mockResolvedValueOnce([{}]); // Mark new versionedTemplate as default
+    tSpy.mockResolvedValueOnce([{}]); // Unmark old template as default
+    vtSpy.mockResolvedValueOnce(undefined); // Unmark old versionedTemplate as default (FAIL)
+    tSpy.mockResolvedValueOnce([{}]); // Mark old template as default (Rollback)
+    vtSpy.mockResolvedValueOnce([{}]); // Mark old versionedTemplate as default (Rollback)
+    tSpy.mockResolvedValueOnce([{}]); // Unmark new template as default (Rollback)
+    vtSpy.mockResolvedValueOnce([{}]); // Unmark new versionedTemplate as default (Rollback)
+
+    await setDefaultTemplate('Test', context, template);
+    expect(Template.query).toHaveBeenCalledTimes(4);
+    expect(VersionedTemplate.query).toHaveBeenCalledTimes(4);
+    expect(Template.query).toHaveBeenLastCalledWith(context, 'UPDATE templates SET isDefault = ? WHERE id = ?;', ['0', template.id.toString()], 'Test');
+    expect(VersionedTemplate.query).toHaveBeenLastCalledWith(context, 'UPDATE versionedTemplates SET isDefault = ? WHERE templateId = ?;', ['0', template.id.toString()], 'Test');
   });
 });